requirements.txt
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cx-Oracle==8.0.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
MarkupSafe==1.1.1
mypy==0.782
mypy-extensions==0.4.3
numpy==1.19.1
pandas==1.1.1
pycodestyle==2.6.0
pylogbeat==2.0.0
python-dateutil==2.8.1
python-logstash-async==2.1.1
pytz==2020.1
requests==2.25.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
xlrd==1.2.0
